apps/requirements.txt
-r ../bm2_python/requirements.txt
paho-mqtt~=1.6.1
